requirements.txt
flask==3.0.0
flask-cors==4.0.1
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.1
pandas==2.0.0
pandas-ta==0.3.14b0
